backend/requirements.txt
newspaper3k
transformers 
torch
googletrans==3.1.0a0
nltk
lxml_html_clean